Abstraction for Custom Cards, Naming, Events
