+++ materialList/src/main/java/com/dexafree/materialList/cards/view/BasicImageButtonsCardItemView.java
@@ -1,0 +1,20 @@
+package com.dexafree.materialList.cards.view;
+
+import android.content.Context;
+import android.util.AttributeSet;
+
+import com.dexafree.materialList.cards.model.BasicImageButtonsCard;
+
+public class BasicImageButtonsCardItemView extends BaseButtonsCardItemView<BasicImageButtonsCard> {
+    public BasicImageButtonsCardItemView(Context context) {
+        super(context);
+    }
+
+    public BasicImageButtonsCardItemView(Context context, AttributeSet attrs) {
+        super(context, attrs);
+    }
+
+    public BasicImageButtonsCardItemView(Context context, AttributeSet attrs, int defStyle) {
+        super(context, attrs, defStyle);
+    }
+}
