+++ materialList/src/main/java/com/dexafree/materialList/cards/view/BigImageButtonsCardItemView.java
@@ -1,0 +1,20 @@
+package com.dexafree.materialList.cards.view;
+
+import android.content.Context;
+import android.util.AttributeSet;
+
+import com.dexafree.materialList.cards.model.BigImageButtonsCard;
+
+public class BigImageButtonsCardItemView extends BaseButtonsCardItemView<BigImageButtonsCard>{
+    public BigImageButtonsCardItemView(Context context) {
+        super(context);
+    }
+
+    public BigImageButtonsCardItemView(Context context, AttributeSet attrs) {
+        super(context, attrs);
+    }
+
+    public BigImageButtonsCardItemView(Context context, AttributeSet attrs, int defStyle) {
+        super(context, attrs, defStyle);
+    }
+}
