+++ materialList/src/main/java/com/dexafree/materialList/events/DataSetChangedEvent.java
@@ -1,0 +1,7 @@
+package com.dexafree.materialList.events;
+
+/**
+ * Created by Fabio on 09.01.2015.
+ */
+public class DataSetChangedEvent {
+}
