+++ materialList/src/main/java/com/dexafree/materialList/view/MaterialListView.java
@@ -9,12 +9,13 @@ import android.view.View;
 import android.widget.BaseAdapter;
 import android.widget.ListView;
 
-import com.dexafree.materialList.MaterialListViewAdapter;
+import com.dexafree.materialList.controller.MaterialListViewAdapter;
 import com.dexafree.materialList.controller.OnDismissCallback;
 import com.dexafree.materialList.controller.SwipeDismissListener;
 import com.dexafree.materialList.events.BusProvider;
+import com.dexafree.materialList.events.DataSetChangedEvent;
 import com.dexafree.materialList.events.DismissEvent;
-import com.dexafree.materialList.model.Card;
+import com.dexafree.materialList.cards.model.Card;
 import com.nhaarman.listviewanimations.appearance.AnimationAdapter;
 import com.nhaarman.listviewanimations.appearance.simple.AlphaInAnimationAdapter;
 import com.nhaarman.listviewanimations.appearance.simple.ScaleInAnimationAdapter;
@@ -150,6 +151,11 @@ public class MaterialListView extends ListView implements IMaterialView {
     }
 
 	@Subscribe
+	public void onNotifyDataSetChanged(DataSetChangedEvent event) {
+		mAdapter.notifyDataSetChanged();
+	}
+
+	@Subscribe
 	public void onCardDismiss(DismissEvent event){
 		Card dismissedCard = event.getDismissedCard();
 		View dismissedCardView = null;
@@ -168,12 +174,12 @@ public class MaterialListView extends ListView implements IMaterialView {
     @Override
     protected void onAttachedToWindow() {
         super.onAttachedToWindow();
-        BusProvider.getInstance().register(this);
+        BusProvider.register(this);
     }
 
     @Override
     protected void onDetachedFromWindow() {
         super.onDetachedFromWindow();
-        BusProvider.getInstance().unregister(this);
+        BusProvider.unregister(this);
     }
 }
